backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
scenedetect[opencv]
openai-whisper
ultralytics
numpy
torch
torchvision
torchaudio
yt-dlp
